Juste réécriture en plusieurs fichiers
Je vais encore réécrire le code en essayant d'utiliser des classes.
Et ça fix #18
